Add vote encrypted ballot
