+++ belenios-api/src/lib/belenios/voter/vote.js
@@ -3,32 +3,7 @@ import { exec } from 'child_process';
 import path from 'path';
 import { ELECTIONS_DIR, BALLOTS_FILE_NAME } from '../global';
 import log from '../../../log';
-
-function saveVote(voteBallot, ballotFilePath) {
-  const publicKey = voteBallot.signature.public_key;
-
-  if (!fs.existsSync(ballotFilePath)) {
-    fs.writeFileSync(ballotFilePath, '');
-  }
-
-  if (publicKey) {
-    const ballots = fs.readFileSync(ballotFilePath);
-
-    if (ballots.includes(publicKey)) {
-      const newBallots = ballots
-        .toString()
-        .split('\n')
-        .map((line) => (line && line.includes(publicKey) ? JSON.stringify(voteBallot) : line))
-        .join('\n');
-      fs.writeFileSync(ballotFilePath, newBallots);
-    } else {
-      fs.appendFileSync(ballotFilePath, `${JSON.stringify(voteBallot)}\n`);
-    }
-
-    return true;
-  }
-  return false;
-}
+import saveVote from './saveVote';
 
 function executeVote(privCred, ballot, ballotFilePath, electionDir, callback) {
   exec(`bash src/scripts/vote.sh ${privCred} ${ballot} ${electionDir}`, (error, stdout) => {
